api/v2, customer: move main class types
